Refactor name-tag to it's own component
